adopt code for checkpoint
